add cbt.nuget.aggregate.unittests and rename MSBuildProjectBuilder.UnitTest to UnitTests
