+++ src/CBT.Modules.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26403.7
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CBT.NuGet", "CBT.NuGet", "{105F8F9A-A30C-4516-9359-916B2D00C29E}"
 EndProject
@@ -51,7 +51,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MSBuildProjectBuilder", "MS
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MSBuildProjectBuilder", "MSBuildProjectBuilder\MSBuildProjectBuilder.csproj", "{69A17EA2-C5F4-44B4-AE85-95018891CDD5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MSBuildProjectBuilder.UnitTest", "MSBuildProjectBuilder.UnitTest\MSBuildProjectBuilder.UnitTest.csproj", "{693000B3-F0AD-4AC7-A06F-DD99BE50220A}"
+Project("{FF286327-C783-4F7A-AB73-9BCBAD0D4460}") = "CBT.NuGet.AggregatePackage", "CBT.NuGet.AggregatePackage\CBT.NuGet.AggregatePackage.nuproj", "{C0FEDA7D-30A2-4E62-BB92-98700DE46B79}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CBT.Nuget.AggregatePackage.UnitTests", "CBT.Nuget.AggregatePackage.UnitTests\CBT.Nuget.AggregatePackage.UnitTests.csproj", "{5869575A-94F7-4880-B6E9-F7F6178D65A8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MSBuildProjectBuilder.UnitTests", "MSBuildProjectBuilder.UnitTests\MSBuildProjectBuilder.UnitTests.csproj", "{693000B3-F0AD-4AC7-A06F-DD99BE50220A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -115,6 +119,14 @@ Global
 		{69A17EA2-C5F4-44B4-AE85-95018891CDD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{69A17EA2-C5F4-44B4-AE85-95018891CDD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{69A17EA2-C5F4-44B4-AE85-95018891CDD5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C0FEDA7D-30A2-4E62-BB92-98700DE46B79}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C0FEDA7D-30A2-4E62-BB92-98700DE46B79}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C0FEDA7D-30A2-4E62-BB92-98700DE46B79}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C0FEDA7D-30A2-4E62-BB92-98700DE46B79}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5869575A-94F7-4880-B6E9-F7F6178D65A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5869575A-94F7-4880-B6E9-F7F6178D65A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5869575A-94F7-4880-B6E9-F7F6178D65A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5869575A-94F7-4880-B6E9-F7F6178D65A8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{693000B3-F0AD-4AC7-A06F-DD99BE50220A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{693000B3-F0AD-4AC7-A06F-DD99BE50220A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{693000B3-F0AD-4AC7-A06F-DD99BE50220A}.Release|Any CPU.ActiveCfg = Release|Any CPU
