+++ src/CBT.NuGet.AggregatePackage/CBT.NuGet.AggregatePackage.nuproj
@@ -3,7 +3,7 @@
   <PropertyGroup Label="Configuration">
     <Description>Provides ability to aggregate nuget packages.</Description>
     <NoPackageAnalysis>true</NoPackageAnalysis>
-    <ProjectGuid>525e50eb-c6d9-497d-bc54-c625ab62b7d0</ProjectGuid>
+    <ProjectGuid>c0feda7d-30a2-4e62-bb92-98700de46b79</ProjectGuid>
     <Tags>CBT Module NuGet Aggregate Packages</Tags>
     <Title>CBT NuGet Aggregate Package Restore</Title>
   </PropertyGroup>
