+++ src/MSBuildProjectBuilder.UnitTests/MSBuildProjectBuilder.UnitTests.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{693000B3-F0AD-4AC7-A06F-DD99BE50220A}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>MSBuildProjectBuilder.UnitTest</RootNamespace>
-    <AssemblyName>MSBuildProjectBuilder.UnitTest</AssemblyName>
+    <RootNamespace>MSBuildProjectBuilder.UnitTests</RootNamespace>
+    <AssemblyName>MSBuildProjectBuilder.UnitTests</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
