+++ src/MSBuildProjectBuilder.UnitTests/PropertyGroupTest.cs
@@ -1,35 +1,35 @@
 ﻿using Microsoft.MSBuildProjectBuilder;
 using Xunit;
 
-namespace MSBuildProjectBuilder.UnitTest
+namespace MSBuildProjectBuilder.UnitTests
 {
-    public class ItemGroupTest
+    public class PropertyGroupTest
     {
 
         [Fact]
-        public void AddItemGroup()
+        public void AddPropertyGroup()
         {
             string expectedOutput =
 @"<?xml version=""1.0"" encoding=""utf-16""?>
 <Project ToolsVersion=""4.0"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <ItemGroup Label=""test label"" Condition=""test Condition"" />
+  <PropertyGroup Label=""test label"" Condition=""test Condition"" />
 </Project>";
             ProjectBuilder.Create()
-                .AddItemGroup("test Condition", "test label")
+                .AddPropertyGroup("test Condition", "test label")
                 .ProjectRoot
                 .RawXmlShouldBe(expectedOutput);
 
             expectedOutput =
 @"<?xml version=""1.0"" encoding=""utf-16""?>
 <Project ToolsVersion=""4.0"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <ItemGroup />
-  <ItemGroup Label=""test label"" Condition=""test Condition"" />
-  <ItemGroup Condition=""New Condition"" />
+  <PropertyGroup />
+  <PropertyGroup Label=""test label"" Condition=""test Condition"" />
+  <PropertyGroup Condition=""New Condition"" />
 </Project>";
             ProjectBuilder.Create()
-                .AddItemGroup()
-                .AddItemGroup("test Condition", "test label")
-                .AddItemGroup("New Condition")
+                .AddPropertyGroup()
+                .AddPropertyGroup("test Condition", "test label")
+                .AddPropertyGroup("New Condition")
                 .ProjectRoot
                 .RawXmlShouldBe(expectedOutput);
         }
