+++ src/MSBuildProjectBuilder.UnitTests/project.json
